Updating new tree sanity tests
This change adjusts the steps performed for the sanity test.  Notably,
it causes the system to be subscribed to 'production' and then upgraded
to the latest tree available.  (This makes the assumption that the new
tree that is being tested is in the 'stage' environment.)

Additionally, the list of RPMs is collected after upgrading to the
latest 'released' tree and then after the system has been updated to the
'stage' tree.  This allows us to produce a diff of the list for later
use.

Finally, the necessary changes to the `steps/rhelah.py` file were made
to support the collection and retrieval of the RPM list.

Signed-off-by: Micah Abbott <miabbott@redhat.com>